Godkjennutpeking og avslaautpeking trenger ikke ligge i felles stegMap-mappe
