+++ src/felleskomponenter/stegvelger/stegMap/index.js
@@ -1,11 +1,7 @@
 import Inngang from './inngang';
 import Virksomheter from './virksomheter';
-import AvslaaUtpeking from './avslaautpeking';
-import GodkjennUtpeking from './godkjennutpeking';
 
 export {
-  AvslaaUtpeking,
-  GodkjennUtpeking,
   Inngang,
   Virksomheter,
 };
