+++ src/sider/vurderutpeking/stegMap/avslaautpeking.js
@@ -1,6 +1,30 @@
-import { AvslaaUtpeking } from '../../../felleskomponenter/stegvelger/stegMap';
+import Steg from '../../../felleskomponenter/stegvelger/stegMotor/steg';
+import { STEG, FANE_STATUS } from '../../../felleskomponenter/stegvelger/stegMotor/typer';
+import VurderingAvslaaUtpeking from '../../../felleskomponenter/stegvelger/stegKomponenter/vurderingAvslaaUtpeking';
 
-class VurderUtpekingAvslaa extends AvslaaUtpeking {
+class AvslaaUtpeking extends Steg {
+  constructor(propsLight, stegPosisjon) {
+    super(propsLight, stegPosisjon);
+
+    this.kriterier = [];
+    this.id = STEG.AVSLAA_UTPEKING;
+    this.tittel = 'Avvis utpeking';
+    this.komponent = VurderingAvslaaUtpeking;
+    this.samleRelevanteData = _propsLight => ({
+      redigerbart: _propsLight.generiskStegRedigerbart,
+    });
+    this.beregnRelevantUI = _propsLight => {
+      const harAvklaring = true;
+
+      return ({
+        harAvklaring,
+      });
+    };
+    this.handlers = {
+      avvisUtpeking: propsLight.tilgjengeligeHandlers.avvisUtpeking,
+    };
+    this._status = FANE_STATUS.OK;
+  }
 }
 
-export default VurderUtpekingAvslaa;
+export default AvslaaUtpeking;
